+++ mrb_doc/models/camera2d.rb
@@ -59,6 +59,25 @@ class Camera2D
   # @return [nil]
   def drawing(&block)
     # src/mruby_integration/models/camera2d.cpp
+    nil
+  end
+
+  # Pass in a vector with world coordinates to see what that translates to in
+  # the camera's viewport.
+  # @params vector [Vector2]
+  # @return [Vector2]
+  def as_in_viewport(vector)
+    # src/mruby_integration/models/camera2d.cpp
+    Vector2.new
+  end
+
+  # Pass in a vector with camera viewport coordinates to see what that
+  # translates to in the world.
+  # @params vector [Vector2]
+  # @return [Vector2]
+  def as_in_world(vector)
+    # src/mruby_integration/models/camera2d.cpp
+    Vector2.new
   end
 end